@@ -13,93 +14,83 @@
   ],
   "exports": {
     ".": {
-      "require": "./dist/bundle/cjs/index.js",
-      "import": "./dist/bundle/esm/index.mjs",
-      "default": "./dist/index.umd.js",
-      "types": "./.build/types/index.d.ts"
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      },
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      }
     },
     "./canvas-panel": {
-      "require": "./dist/canvas-panel/cjs/canvas-panel.js",
-      "import": "./dist/canvas-panel/esm/canvas-panel.mjs",
-      "types": "./.build/types/canvas-panel/index.d.ts"
+      "require": {
+        "types": "./dist/canvas-panel.d.cts",
+        "default": "./dist/canvas-panel.cjs"
+      },
+      "import": {
+        "types": "./dist/canvas-panel.d.ts",
+        "default": "./dist/canvas-panel.js"
+      }
     },
     "./utils": {
-      "require": "./dist/utils/cjs/utils.js",
-      "import": "./dist/utils/esm/utils.mjs",
-      "types": "./.build/types/utils.d.ts"
-    },
-    "./react17": {
-      "require": "./dist/react17/cjs/index.js",
-      "import": "./dist/react17/esm/index.mjs",
-      "types": "./.build/types/index.d.ts"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "*": [
-        ".build/types/index.d.ts"
-      ],
-      "utils": [
-        ".build/types/utils.d.ts"
-      ],
-      "canvas-panel": [
-        ".build/types/canvas-panel/index.d.ts"
-      ]
+      "require": {
+        "types": "./dist/utils.d.cts",
+        "default": "./dist/utils.cjs"
+      },
+      "import": {
+        "types": "./dist/utils.d.ts",
+        "default": "./dist/utils.js"
+      }
     }
   },
   "scripts": {
-    "start": "vite",
-    "build": "tsc -p ./tsconfig.types.json --declaration --emitDeclarationOnly && node vite/build.mjs",
-    "prepublishOnly": "tsc -p ./tsconfig.types.json --declaration --emitDeclarationOnly && node vite/build.mjs",
-    "test": "vitest"
+    "build": "tsup && tsup --config tsup.umd.ts",
+    "dev": "tsup --watch",
+    "typecheck": "tsc --noEmit",
+    "test": "vitest",
+    "prepack": "tsup && tsup --config tsup.umd.ts",
+    "lint": "publint",
+    "start": "vite"
   },
   "license": "MIT",
   "resolutions": {
-    "@atlas-viewer/iiif-image-api": "https://pkg.csb.dev/atlas-viewer/iiif-image-api/commit/7d4a2fe7/@atlas-viewer/iiif-image-api",
-    "react-reconciler": "^0.28",
-    "@iiif/presentation-3": ">=2.2.2"
+    "react-reconciler": "0.29"
   },
   "dependencies": {
-    "@atlas-viewer/atlas": ">=2.1.0",
-    "@atlas-viewer/iiif-image-api": "https://pkg.csb.dev/atlas-viewer/iiif-image-api/commit/7d4a2fe7/@atlas-viewer/iiif-image-api",
-    "@iiif/helpers": ">=1.0.6",
-    "@iiif/parser": ">=2.0.2",
-    "@iiif/presentation-2": ">=1.*",
-    "@iiif/presentation-3": ">=2.2.2",
-    "@iiif/presentation-3-normalized": ">=0.9.7",
-    "react": "^16.10.2 || ^17.0.2 || ^18.2.0",
-    "react-dom": "^16.10.2 || ^17.0.2 || ^18.2.0",
-    "react-reconciler": "^0.28",
-    "react-error-boundary": "^4.0.12",
-    "react-lazy-load-image-component": "^1.6.0"
+    "@atlas-viewer/atlas": "^2.2.2",
+    "@atlas-viewer/iiif-image-api": "^2.2.0",
+    "@iiif/helpers": "^1.1.0",
+    "@iiif/parser": "^2.1.0",
+    "@iiif/presentation-2": "^1.0.4",
+    "@iiif/presentation-3": "^2.2.2",
+    "@iiif/presentation-3-normalized": "^0.9.7",
+    "react-error-boundary": "^4.0.13",
+    "react-lazy-load-image-component": "^1.6.0",
+    "react-reconciler": "~0.29.0",
+    "zustand": "^4.5.2"
+  },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@testing-library/react": "^14.2.1",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@types/react": "^18.2.55",
-    "@types/react-dom": "^18.2.19",
+    "@testing-library/react": "^14.3.0",
+    "@types/react": "^18.2.75",
+    "@types/react-dom": "^18.2.24",
     "@types/react-lazy-load-image-component": "^1.6.3",
-    "@typescript-eslint/eslint-plugin": "^7.0.1",
-    "@typescript-eslint/parser": "7.0.1",
     "@vitejs/plugin-react": "^4.2.1",
-    "chalk": "^5.3.0",
-    "eslint": "^8.56.0",
-    "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "happy-dom": "^13.3.8",
+    "esbuild-plugin-globals": "^0.2.0",
+    "esbuild-plugin-umd-wrapper": "^2.0.0",
+    "happy-dom": "^14.7.1",
     "prettier": "^3.2.5",
-    "query-string": "^8.2.0",
-    "react-16": "npm:react@16",
-    "react-17": "npm:react@17",
-    "react-18": "npm:react@18",
-    "react-dom-16": "npm:react-dom@16",
-    "react-dom-17": "npm:react-dom@17",
-    "react-dom-18": "npm:react-dom@18",
+    "publint": "^0.2.7",
+    "query-string": "^9.0.0",
     "tslib": "^2.6.2",
-    "typescript": "^5.3.3",
-    "vite": "^5.1.2",
-    "vitest": "^1.2.2",
-    "zustand": "4.5.2"
+    "tsup": "^8.0.2",
+    "typescript": "^5.4.4",
+    "vite": "^5.2.8",
+    "vitest": "^1.4.0"
   },
   "publishConfig": {
     "access": "public"
